review fixes for hw-1
