+++ src/main/java/com/book/booksstore/repository/BookRepository.java
@@ -5,5 +5,6 @@ import java.util.List;
 
 public interface BookRepository {
     Book save(Book book);
+
     List<Book> findAll();
 }
