+++ src/main/java/com/book/booksstore/repository/BookRepositoryImpl.java
@@ -3,7 +3,9 @@ package com.book.booksstore.repository;
 
 import com.book.booksstore.model.Book;
 import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
 import java.util.List;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Repository;
 import jakarta.persistence.PersistenceContext;
 
@@ -14,17 +16,24 @@ public class BookRepositoryImpl implements BookRepository  {
     private EntityManager entityManager;
 
     @Override
+    @Transactional
     public Book save(Book book) {
-        if (book.getId() == null) {
-            entityManager.persist(book);
-            return book;
-        } else {
-            return entityManager.merge(book);
+        try {
+            if (book.getId() == null) {
+                entityManager.persist(book);
+                return book;
+            } else {
+                return entityManager.merge(book);
+            }
+        } catch (Exception e) {
+            // You can wrap it in a custom exception or just rethrow
+            throw new DataAccessException("Failed to save book: " + book, e) {};
         }
+
     }
 
     @Override
-    public List findAll() {
+    public List<Book> findAll() {
         return entityManager.createQuery("SELECT b FROM Book b", Book.class).getResultList();
     }
 }
