--- src/main/java/com/book/booksstore/service/BookServiceInterface.java
@@ -1,10 +1,0 @@
-package com.book.booksstore.service;
-
-import com.book.booksstore.model.Book;
-import java.util.List;
-
-public interface BookServiceInterface {
-    Book save(Book nextBook);
-
-    List findAll();
-}
