+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.application.name=books-store
-spring.datasource.url=jdbc:mysql://localhost:3306/books
+spring.datasource.url=jdbc:mysql://localhost:3306/booksstore
 spring.datasource.username=root
 spring.datasource.password=rootroot
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
